Move TCP & UDP types into `net` module.
Motivation for this is twofold.

1) Mirror the layout in `std`
2) Fix UDP API without breaking backwards compatibility.
