+++ src/sys/unix/udp.rs
@@ -1,6 +1,5 @@
 use {io, Ready, Poll, PollOpt, Token};
 use event::Evented;
-use io::MapNonBlock;
 use unix::EventedFd;
 use std::net::{self, Ipv4Addr, Ipv6Addr, SocketAddr};
 use std::os::unix::io::{RawFd, IntoRawFd, AsRawFd, FromRawFd};
@@ -33,31 +32,23 @@ impl UdpSocket {
         })
     }
 
-    pub fn send_to(&self, buf: &[u8], target: &SocketAddr)
-                   -> io::Result<Option<usize>> {
+    pub fn send_to(&self, buf: &[u8], target: &SocketAddr) -> io::Result<usize> {
         self.io.send_to(buf, target)
-            .map_non_block()
     }
 
-    pub fn recv_from(&self, buf: &mut [u8])
-                     -> io::Result<Option<(usize, SocketAddr)>> {
+    pub fn recv_from(&self, buf: &mut [u8]) -> io::Result<(usize, SocketAddr)> {
         self.io.recv_from(buf)
-            .map_non_block()
     }
 
-    pub fn send(&self, buf: &[u8])
-                   -> io::Result<Option<usize>> {
+    pub fn send(&self, buf: &[u8]) -> io::Result<usize> {
         self.io.send(buf)
-            .map_non_block()
     }
 
-    pub fn recv(&self, buf: &mut [u8])
-                     -> io::Result<Option<usize>> {
+    pub fn recv(&self, buf: &mut [u8]) -> io::Result<usize> {
         self.io.recv(buf)
-            .map_non_block()
     }
 
-    pub fn connect(&self, addr: SocketAddr) 
+    pub fn connect(&self, addr: SocketAddr)
                      -> io::Result<()> {
         self.io.connect(addr)
     }
